Marginally improve code generation.
We reserve all attributes beginning with _ and __ as part of the implementation, and discourage (i.e. prevent, but there are probably workarounds) user-written code from accessing or modifying these.
Also rename self.marked to self._marked and self.__successors to self._successors, and prevent user-defined attributes/outputs being called this.